refactor simpler glob match with one wildcard only
