+++ capability/policy/match.go
@@ -181,69 +181,3 @@ func gt(order int) bool  { return order == 1 }
 func gte(order int) bool { return order == 0 || order == 1 }
 func lt(order int) bool  { return order == -1 }
 func lte(order int) bool { return order == 0 || order == -1 }
-
-// globMatch matches a string against a pattern with '*' and '?' wildcards, handling escape sequences.
-func globMatch(pattern, str string) bool {
-	var i, j int
-	for i < len(pattern) && j < len(str) {
-		switch pattern[i] {
-		case '*':
-			// skip consecutive '*' characters
-			for i < len(pattern) && pattern[i] == '*' {
-				i++
-			}
-			if i == len(pattern) {
-				return true
-			}
-
-			// match the rest of the pattern
-			for j < len(str) {
-				if globMatch(pattern[i:], str[j:]) {
-					return true
-				}
-				j++
-			}
-
-			return false
-		case '?':
-			// match any single character
-			i++
-			j++
-		case '\\':
-			// Handle escape sequences
-			i++
-			if i < len(pattern) && pattern[i] == '*' {
-				if str[j] != '*' {
-					return false
-				}
-				i++
-				j++
-			} else if i < len(pattern) && pattern[i] == '?' {
-				if str[j] != '?' {
-					return false
-				}
-				i++
-				j++
-			} else {
-				if i >= len(pattern) || pattern[i] != str[j] {
-					return false
-				}
-				i++
-				j++
-			}
-		default:
-			if pattern[i] != str[j] {
-				return false
-			}
-			i++
-			j++
-		}
-	}
-
-	// check for remaining characters in pattern
-	for i < len(pattern) && pattern[i] == '*' {
-		i++
-	}
-
-	return i == len(pattern) && j == len(str)
-}
